Move test_secret_removed_from_show_techsupport to new file because UT in test_techsupport.py can't run with KVM.
